fix: Launcher update only work for launcher update
